Move tests to web app (#136)
Since indexiing is now part of the web app, the tests have to
be placed there, too.
