+++ web/build.sbt
@@ -11,5 +11,6 @@ libraryDependencies ++= Seq(
   javaWs,
   "org.elasticsearch" % "elasticsearch" % "2.3.1"
       // otherwise javaWs won't work
-      exclude ("io.netty", "netty")
+      exclude ("io.netty", "netty"),
+  "com.github.jsonld-java" % "jsonld-java" % "0.5.0"
 )
