+++ web/test/controllers/ElasticsearchTest.java
@@ -1,7 +1,8 @@
-package flow;
+package controllers;
 
 import static org.elasticsearch.node.NodeBuilder.nodeBuilder;
 
+import java.io.File;
 import java.io.IOException;
 
 import org.elasticsearch.action.search.SearchResponse;
@@ -13,31 +14,31 @@ import org.elasticsearch.node.Node;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 
+import com.typesafe.config.Config;
+import com.typesafe.config.ConfigFactory;
+
 @SuppressWarnings("javadoc")
 public abstract class ElasticsearchTest {
 
-	protected static Node node;
-	protected static Client client;
+	public static final Config CONFIG =
+			ConfigFactory.parseFile(new File("conf/application.conf")).resolve();
+
+	protected static Client client = Index.CLIENT;
 	protected static Node geoNode;
 	protected static Client geoClient;
 
 	@BeforeClass
 	public static void makeIndex() throws IOException {
-		node = nodeBuilder().settings(Settings.builder().put("path.home", "."))
-				.local(true).node();
-		client = node.client();
 		setupLocalGeodataExample();
-		transformData();
-		prepareIndexing(client, Constants.ES_INDEX,
-				Constants.MAIN_RESOURCES_PATH + "index-settings.json");
-		indexData(client, Constants.TEST_RESOURCES_PATH + Constants.OUTPUT_PATH
-				+ "enriched.out.json", Constants.ES_INDEX);
+		prepareIndexing(client, CONFIG.getString("index.es.name"),
+				"conf/index-settings.json");
+		indexData(client, "resources/enriched-test.json",
+				CONFIG.getString("index.es.name"));
 	}
 
 	@AfterClass
 	public static void closeElasticSearch() {
 		client.close();
-		node.close();
 		geoClient.close();
 		geoNode.close();
 	}
@@ -48,13 +49,7 @@ public abstract class ElasticsearchTest {
 		geoClient = geoNode.client();
 		String geoIndex = "geodata";
 		prepareIndexing(geoClient, geoIndex, null);
-		indexData(geoClient, Constants.TEST_RESOURCES_PATH + Constants.INPUT_PATH
-				+ "AA603_geodata.json", geoIndex);
-	}
-
-	public static void transformData() throws IOException {
-		EnrichSample.processSample(Constants.TEST_RESOURCES_PATH
-				+ Constants.OUTPUT_PATH + "enriched.out.json");
+		indexData(geoClient, "resources/geodata-test.json", geoIndex);
 	}
 
 	public static void prepareIndexing(final Client aIndexClient,
@@ -71,7 +66,8 @@ public abstract class ElasticsearchTest {
 	public static SearchResponse exactSearch(final String aField,
 			final String aValue) {
 		final SearchResponse responseOfSearch =
-				client.prepareSearch(Constants.ES_INDEX).setTypes(Constants.ES_TYPE)
+				client.prepareSearch(CONFIG.getString("index.es.name"))
+						.setTypes(CONFIG.getString("index.es.type"))
 						.setSearchType(SearchType.DFS_QUERY_AND_FETCH)
 						.setQuery(QueryBuilders.termQuery(aField, aValue)).execute()
 						.actionGet();
@@ -81,7 +77,8 @@ public abstract class ElasticsearchTest {
 	public static SearchResponse search(final String aField,
 			final String aValue) {
 		SearchResponse responseOfSearch =
-				client.prepareSearch(Constants.ES_INDEX).setTypes(Constants.ES_TYPE)
+				client.prepareSearch(CONFIG.getString("index.es.name"))
+						.setTypes(CONFIG.getString("index.es.type"))
 						.setSearchType(SearchType.DFS_QUERY_AND_FETCH)
 						.setQuery(QueryBuilders.matchQuery(aField, aValue)).execute()
 						.actionGet();
