+++ web/test/controllers/TestJsonStructure.java
@@ -1,4 +1,5 @@
-package flow;
+package controllers;
+
 
 import static org.junit.Assert.assertEquals;
